Adds model and integration tests for Entry.es_search and Entry.filtered_search; updates Entry.filter to take options hash

+++ spec/support/helpers/entry_support.rb
@@ -1,21 +1,24 @@
 # spec/support/helpers/entry_support.rb
 module EntrySupport
-  def each_entry_type(user, friend, visitor)
+  # Create one of each of the 6 possible entry types from permutations of
+  # private/public and visitor/friend/orphan.
+  def each_entry_type(user, friend, visitor, opts = {})
+    opts.delete_if { |_key, val| val.blank? }
     {
-      user_ent_1: create(:entry, user: user,   public: true,  visitor: visitor),
-      user_ent_2: create(:entry, user: user,   public: false, visitor: visitor),
-      publ_orph:  create(:entry, user: nil,    public: true,  visitor: nil),
-      priv_orph:  create(:entry, user: nil,    public: false, visitor: nil),
-      publ_ent:   create(:entry, user: friend, public: true,  visitor: nil),
-      priv_ent:   create(:entry, user: friend, public: false, visitor: nil)
+      user_ent_1: create(:entry, opts.merge(user: user,   public: true,  visitor: visitor)),
+      user_ent_2: create(:entry, opts.merge(user: user,   public: false, visitor: visitor)),
+      publ_orph:  create(:entry, opts.merge(user: nil,    public: true,  visitor: nil)),
+      priv_orph:  create(:entry, opts.merge(user: nil,    public: false, visitor: nil)),
+      publ_ent:   create(:entry, opts.merge(user: friend, public: true,  visitor: nil)),
+      priv_ent:   create(:entry, opts.merge(user: friend, public: false, visitor: nil))
     }
   end
 
-  def create_dummy_entries
-    Entry.delete_all
+  def create_dummy_entries(clear: true, title: nil)
+    Entry.delete_all if clear  # Clear all existing entries by default.
     @user ||= create(:user)
     @visitor ||= create(:visitor, user: @user)
     @friend ||= create(:user)
-    @entries = each_entry_type(@user, @friend, @visitor)
+    @entries = each_entry_type(@user, @friend, @visitor, title: title)
   end
 end
